requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
backports.pbkdf2==0.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
commonmark==0.9.1
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.4
Django==4.1.2
django-camunda==0.11.1
django-crispy-forms==1.14.0
django-solo==2.0.0
djangorestframework==3.14.0
filelock==3.6.0
h11==0.13.0
hachoir==3.1.3
httpie==3.2.1
idna==3.3
inflection==0.5.1
lxml==4.9.1
mpi4py==3.1.3
multidict==6.0.2
numpy==1.23.0
outcome==1.2.0
pandas==1.4.3
Pillow==9.3.0
platformdirs==2.5.1
protobuf==4.21.9
psycopg2==2.9.3
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.5
requests==2.27.1
requests-toolbelt==0.10.1
rich==12.6.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
virtualenv==20.13.4
whitenoise==6.2.0
wsproto==1.1.0
